feat: added any models; implemented 30 percent of project; implemented chase, shoot, health, visiblityField, kickPlayer, enemyAnimation, shootAnimation mechanic; changed main scene

+++ Assets/Scripts/Core/Events/PlayerHealthChanged.cs
@@ -2,9 +2,9 @@
 {
     public struct PlayerHealthChanged
     {
-        public readonly int Health;
+        public readonly float Health;
 
-        public PlayerHealthChanged(int health)
+        public PlayerHealthChanged(float health)
         {
             Health = health;
         }
